Update README.md
PROBLEM STATEMENT:

We need to implement a solution for sending mass mailings using AWS Lambda. The goal is
to create a scalable and efficient system that can handle the distribution of a large number of
emails to a recipient list while ensuring high deliverability and minimal operational overhead.
This project should be implemented with a focus on security, reliability, and cost-effectiveness.